[FIX] appointment{_account_payment}: ensure paid resource capacity_used
ISSUE
=====

Introduced in odoo/enterprise@bce7e94650c337a9046958a7689c81db5b2a4c73 ,
we now allow configuring a number of max simulatenous appointments per
resource / user per slot, when manage_capacity is False.

However, this broke the booking flow when enabling a payment on the
appointment type as well, for resources with capacity > 1. More
specifically, as soon as the remaining capacity is lower than the
resource's own capacity.

STEPS TO REPRODUCE
==================

0. Install payment_demo, in order to be able to test the payment issue.
Install website_appointment_sale to ease the testing too.
1. Create an appointment type for 'Resources', enable 'Up-front payment'.
2. Set a new resource with capacity = 2 on the appointment type.
3. Make sure manage_capacity is False, and number of simultaneous
appointments (max_bookings) is set to 1.
4. Go to the front end and pick any slot.
5. Fill the form and 'proceed to payment'
6. Try to 'pay now'
7. You get a validation error message telling you that the slot is not
available anymore. But it is not the case since it is a new type and new
resource.

FIX
===

For resources, created booking lines were setting capacity_used to the
resource's capacity, unless shared. (meaning: the resource is completely
used, even if the reserved quantity is smaller). It made sense before but
not for this new feature. It should only use 1 capacity, as we count the
number of different simulatenous appointments, and it is unrelated to the
resource's own capacity.

When creating the meeting at the end of the regular flow, the computed field
on capacity_used on appointment.booking.line will set it to 1 when
manage_capacity is False. Hence, no issue will happen there in practice as the
meeting is always created in that case and capacity_used updated.

However, when enabling a payment, we create a calendar.booking instead, with
calendar.booking.lines. And based on those, we evaluate availability when
checking the cart before going to payment.

Those are simply created from values in the controller, and the model is much
simpler. Therefore, we fix the issue by manually setting the capacity_used to
1 when creating the booking lines in the front-end booking flow, in the case
of an appointment without capacity management.

Task-4963012

closes odoo/enterprise#94434

X-original-commit: c3eb8e79b47573a3023d03650a2922de8860efb4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Noé Antoine (nan) <nan@odoo.com>