[REV] account_asset: consider imported depreciated amount
This reverts commit 79aa48867de14a6d2315ec95b271004084d74b14.

A lot of customers reported that the data was wrong in a lot of cases.
For instance, when the previous software computed a lower depreciation
than Odoo, Odoo was then moving the difference into the next period.

The change for the computation of `asset_remaining_value` and
`asset_depreciated_value` has been kept.

opw-4968367
opw-4977970
opw-4961381
opw-4991216
opw-4965451

closes odoo/enterprise#91780

X-original-commit: 895e14483b0077106749af301b20e490dc97b3c9
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>