[PERF] quality_repair: Speed-up `_create_quality_checks_for_repair`
Description
-----------
Creating a repair creates new `quality.check` based on `quality.
point`s. When you have many matching points to explore, the
`filtered_domain` can be pretty slow when processing linear the
records to compute the intersection between the all points and the
points that should be associated with the repair.

This commit rewrites `_create_quality_checks_for_repair` to use a
`search` instead if necessary, to offload the intersection search in
the database.

We also split the ids to not inject a large list of ids in the query.

Benchmark
---------
For a database that will have 500k `quality.point` that satisfies the
domain of `quality.point._get_domain`, the invocation of
`_create_quality_checks_for_repair` for creating a Repair takes
(excluding the `create` call time):

|         | Before           | After |
|---------|------------------|-------|
| Timings | Timeout (15+min) | 1 min |

Reference
---------
opw-4980036

closes odoo/enterprise#95700

X-original-commit: 28f76a2e2331da44363a04acad13ab37b5ae5abe
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>