[FIX] voip: don't mix server and client values when computing time
It was reported that the timer for the call would sometimes start as a
negative number. Examining the code that computes it suggests that the
timestamp sent by the server was a few seconds in the future compared to
luxon.DateTime.now() (client time). This discrepancy is most likely due
to the clock skew between the client and the server.

This commit adapts the code so that it only computes time based on
client-side values, effectively preventing clock skew issues.

Task-4930666

closes odoo/enterprise#91951

X-original-commit: 35d9d55a2f0bf76b0ab4cf4f62cb7b3d0a07248c
Signed-off-by: Yuchen Huang (yhu) <yhu@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>