[IMP] appointment: iconized gantt form
Replace all the field labels in the calendar quick edit view with
icons, matching changes done in the calendar app. This simply inherits
the calendar quick edit and sets it to be used as the quick create form
from other appointment views (gantt and calendar).

The aim is to keep these views very focused on a few key fields for people
to quickly be able to create a booking, and incite them to go to the full
view to go in more details.

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>