[FIX] l10n_mx_edi: Unrequire tax object for line section and line note
Issue:
When we try to save an invoice with only a line section or line note,
an error is thrown indicating invalid fields: invoice lines.
This happens because the `l10n_mx_edi_tax_object` is required on account move lines.

Purpose of this PR:
Allow for line sections and line notes to be saved without tax objects.

Steps to reproduce on Runbot:
install l10n_mx
switch to MX company and create an invoice with just a line section or line note
save
Invalid Fields error is raised

opw-4987040

closes odoo/enterprise#91792

X-original-commit: af1ec52ec72961dce1dec5c8ae8ce8413f487b0a
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ian Dao (iada) <iada@odoo.com>