[FIX] planning: prevent error when entering invalid end hour in shift template
Currently, an error occurs when entering an invalid end hour in the
shift template.

Steps to Reproduce:
 - Install the planning module.
 - Go to Shift Templates.
 - Create a new shift template and enter an invalid end hour (greater than
   or equal to 24).

ValueError: hour must be in 0..23

This error occurs after [this commit](a0bb61f), which added the _get_name
method to generate the display name without validating the data.
when a user enter invalid value in the end hour while creating or editing
a shift template. The compute method is then triggered, which attempts to
generate the display name [1] based on the formatted time. During the
formatting of the end time, the hour value is invalid for the time function,
and raise the error [2].

This commit ensures that the newly added method _get_name raises a
validation error if the start hour or end hour is invalid.

[1]: https://github.com/odoo/enterprise/blob/3009d64c72ef3f4e6bae64ddb9e24f48a4f778a5/planning/models/planning_template.py#L54

[2]: https://github.com/odoo/enterprise/blob/3009d64c72ef3f4e6bae64ddb9e24f48a4f778a5/planning/models/planning_template.py#L74C1-L79C1

sentry-6792658060

closes odoo/enterprise#92709

X-original-commit: 15c84c9a8ecd4d2b0d9ff169d34b9edef56d9e1c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>