[FIX] pos_enterprise: fix crash when changing order state in preparation display
Currently, an IndexError traceback occurs when changing the order state
in the preparation display if it contains only one stage.

Steps to reproduce this issue:
1) Install POS, Kitchen Display
2) Create a preparation display by removing all but
   one stage in the prep settings.
3) Open a restaurant session and create an order.
4) Open the preparation display and mark the created order as DONE.
5) A traceback will occur

Error:
"IndexError: tuple index out of range"

Cause:
When the Done button is clicked in a preparation display with only one stage,
an ORM call to change_state_status is triggered. This then calls
_record_status_change_prep_time, followed by is_stage_position.

In is_stage_position, static positions [0, -1, -2] are used to access
items in stage_ids. If only one stage exists, accessing indices -1 or -2 results
in an IndexError.

Solution:

Before accessing a stage by position, check that the length of stage_ids is
greater than the absolute value of the position.
This prevents attempts to access out-of-range indices.

opw-4985306

closes odoo/enterprise#92800

X-original-commit: 6df39743d1b4b95543a1381bf5eb75f00c2ba23c
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>