[FIX] documents,*: avoid link_to_record crash with no existing record
*= documents_fleet

Currently, when user tries to link a document with a vehicle and if there is no
vehicle exists an error is raised.

Steps to Reproduce:
1) Install 'documents_fleet' module(without Demo Data).
2) Open Documents App.
3) Upload a file in **Fleet** folder.
4) Select the uploaded file and click on **'Link to a vehicle'** button.

Error:
ValueError: invalid literal for int() with base 10: 'False'

Root Cause:
The 'link_to_record' method attempts to find the first record of
the target model to use as a default value when opening the linking wizard.
If no records exist, the search returns an empty recordset, and accessing '.id'
at [1], on it results in the value 'False'. The code then incorrectly constructs
a default reference string like 'fleet.vehicle,False'. The wizard crashes when
it tries to parse this string and convert the 'False' part to an integer.

Solution:
This commit prevents error by ensuring that user cannot link a document to a
model without existing records.

[1]: https://github.com/odoo/enterprise/blob/83640d21b2c31ed9d572c48e02a9aeea9eb1607b/documents/models/documents_workflow_rule.py#L113

sentry-6792079496

closes odoo/enterprise#93892

X-original-commit: 6ac41577ba3f9587638d3be79a6bb37c0a7aad5e
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Het Patel (path) <path@odoo.com>