[FIX] hr_*_skill: allow including other fields in copy
Allow models that inherit the hr_individual_skill mixin to specify
additional fields to be included/preserved/carried-over when
writing/creating new skill records.
This was already an issue in the `hr_appraisal_skill` module where
the `Justification` field was not included when copying/creating new
skills. As a result, the field was "erased" on every skill level change

task-4984689

closes odoo/enterprise#93275

X-original-commit: 8791af0bc13f9d7e303308cc378841a04c249a25
Related: odoo/odoo#224413
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>