[FIX] l10n_se_sie4_import: Don't import moves without transaction
The aim of this commit is making sure that we don't import move
without transaction in it.

In SIE4, we can have moves with BTRANS and RTRANS tags that we are
ignoring. It means that we could have move without any move lines.
It's not an expected behavior in Odoo. This commit checks that the
move data are created only if we have lines for that move.

no task id

closes odoo/enterprise#92068

X-original-commit: 3032d40aac72a66887e4e8d70521a80769241e33
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>