[FIX] industry_fsm_stock: add map view for outgoing picking
This commit is the continuity of the fix made in community. The issue
fixed in community was the map view cannot be defined in the view mode
of an action since the map view only exists in enterprise.
This commit overrides the method defined in community to add the map
view as before.

Related PR: odoo/odoo#222003

closes odoo/enterprise#92190

X-original-commit: 3a22f3c6cddb0b501ce5319e7def82dd364b6de6
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>