[FIX] account_reports: sort by am.id when fetching AMLs for journal report
Issue description:
When fetching AMLs for the journal report with _query_aml(),
it sorts the AMLs based on (am.date, am.name), which are not unique
in case we have multiple (draft) moves with the same date.
The lines will end up mixed and ordered with respect to the account,
which causes errors with the "Load More" functionality,
as it assumes that the lines are ordered based on their move.
For posted entries, it's not an issue as the am.name is unique.

Steps to reproduce:
-Create 3 or more invoices (with 3+ AMLs each) in draft and on the same invoicing date.
-Open journal report settings and set the Load More Limit to 5.
-Open the journal report and set the date to this day and check
the "Include Draft Entries" option.
-Press "Load More", you will get an error.

opw-4929907

closes odoo/enterprise#91885

X-original-commit: 6c4e027ac62cb892e167606e95fec6a1d740e141
Signed-off-by: Mohamed Yousef Ahmed (myah) <myah@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>