[FIX] account_winbooks_import: skip check on existing CoA
When importing Winbooks data, if the CoA configuration in Winbooks is
different from Odoo's default, users will experience a failure with a
message similar to

```
Validation Error
Tax Receivable Account (411000 VAT Recoverable) should be an account of
type "Receivable" or "Payable" with both options "Allow Reconciliation"
and "Non Trade" enabled.
```

This occurs because the account 411000, meant to store the tax provisions,
in Winbooks was used as tax receivable.
In similar cases we should not raise any error and should let the import
go through: the data is imported in draft and the user has the possibility
to  bring the modifications in Odoo

opw-4850314

closes odoo/enterprise#91955

X-original-commit: 20804db6d87d6613172f4b3ec72277c5a923b335
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>