[FIX] l10n_it_pos: negative payment
Before this commit, specifying the payment method used to give back the change in a POS order would lead to the fiscal printer receiving a negative payment as input, which is not supported, thus leading to the order not being treated by the fiscal printer (so not reported to the government). Only the payments with is_change=true where filtered out.
I am now adding a filter to only keep the positive payments in the receipt, as the fiscal printer is computing the change itself.

opw-4931671

closes odoo/enterprise#93558

X-original-commit: fce4be2faa29efad29f9be7c235704e8a6359faf
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Maxime Orban (orma) <orma@odoo.com>