[REF] ai: follow-up for the transcription component
This commit is a follow-up of the transcription task that improves the
voice transcription component. It adds the features that weren't
implemented by the previous task, as well as some improvements and bug
fixes for existing behaviours.

Among the notable improvements, we have:
- Client-side detection of the audio: the silences in the speech are now
detected from the client. When a silence is detected, the ongoing audio
is committed for transcription. This allows for a more consistent
experience for the user,
- Introduction of default prompts: the component now uses the default
prompts for the summary feature. It allows for a more robust selection
of the agent, and adds action buttons in the summary tab for re-writing
of the summary,
- Introduction of feedbacks: the component now shows a feedback to show
that it is listening to the audio,
- Various bug fixes: some bug fixing of crashes that were experienced,
especially in html fields outsite of knowledge,
- Component deletion: the recording is now properly stopped when
deleting the component in the editor while it is still ongoing,
- Introduction of hoot tests for the transcription component.

task-5002608

closes odoo/enterprise#91797

Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>