[FIX] pos_blackbox_be: fix issue when doing multiple bill on same order
Before this commit an issue was appearing when we try to make multiple bill on the same order. If the order was already synced and didn't get any changes, clicking `Print bill` with `pos_blackbox_be` wouldn't do anything. Now we print the bill after syncing the order, even if the order don't have any new changes to sync.

closes odoo/enterprise#92133

Task-id: 4901299
X-original-commit: ab17afb1d4e66d65d02b45d746da30f912b103dc
Related: odoo/odoo#222537
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Manu Vaillant (manv) <manv@odoo.com>