[FIX] accountant_knowledge: fix error when loading embedded PDF files
Currently, the method responsible for loading PDF files embedded in an
article crashes because it receives an `ir.attachment` object instead
of a bytes-like object (-> TypeError: a bytes-like object is required,
not 'ir.attachment'). Because of this error, it is currently not
possible to export an audit report containing an embedded PDF file.

Steps to reproduce:
1. Create an audit report
2. Open the article associated with the audit report
3. Insert a PDF file in it using the `/file` command
4. Export the article to PDF

=> The page loading the PDF shows an internal error.

To revolve this issue, we will provide the raw attachment to the method
responsible for loading the PDF file. With this fix, people should be
able to export an audit report containing PDF files without errors.

Task-4989809

closes odoo/enterprise#91823

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>