[FIX] approvals: duplicate user for two approvers on same request
Reproduce: Approvals app -> My Approvals -> My Requests (Traceback Error)

Issue: The demo data contains a duplicate record for an approver (the user is set as default approver for the approval category and added again as record in xml data), so the request had 2 approvers with the same user.

Fix: replaced the `@api.constrains` with a UNIQUE SQL constraint to enforce the check of same user linked to more than one approver for single request, and deleted duplicate record from demo xml data.

task-4984137

closes odoo/enterprise#91829

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>