[IMP] appointment{_hr}: add flexible schedule option
PURPOSE
=======

Some users would like to use unique slots instead of having
recurring ones. This can only be done from the share availabilities
button in the calendar view. While the appointment can be modified,
it is not intuitive to do it that way. We want to make the 'custom'
appointments easier to use and accessible.

MAIN CHANGES
============

As 'custom' appointment type means 'flexible' schedule and vice versa,
we do not store any more field. Instead, we use an editable computed
selection field category_slot_scheduling that allows the user to
switch between options with the radio widget, and update category and
slots accordingly in the onchange and inverse (on category) methods.

NEW SLOT CONSTRAINT
===================

Since unique slots are easier to user, let us add a sql constraint to
ensure that the start_datetime and end_datetime are set for the unique
slots, and that start_datetime <= end_datetime.

We make sure we correctly update the slots when changing the category
in the appointment type form (see onchange and inverse methods)

OTHER CHANGES
=============

- Slight design change in the calendar view: remove separate 'Share
Availabilities' button next to the dropdown. Instead, only use the
dropdown button and give it a 'Share' label. Tests are updated.
- Make 'Restrict to work hours' invisible for custom appointments since
it has no effect in that case.
- New and renamed filters and groupby for appointment.type
- Some tests are added to ensure constraint and form reactivity on
category change.

UPGRADE
=======

- We remove the 'invitations' submenu in the configuration that allowed
to see all 'custom' appointments. It is not relevant anymore and is
replaced with an available filter on the appointment.type search view.
We also remove the associated view.

- We remove ill-configured slots. As they are only used to generate
slots and not linked to other records, we can remove them altogether.
(constraints would still make them fail in the booking flow, so they
are not really working anyway)

UPG PR: odoo/upgrade#8206

Task-4997980

Part-of: odoo/enterprise#91824
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>