[IMP] account_accountant: adding docstring
In the backport of: https://github.com/odoo/enterprise/commit/8eda04cdc9a32ceb613026c6928776678adc050b
We took the liberty to add a docstring and a comment to make the code a bit more
readable. This commit will add the same docstring and comment to explain the
get_all_substrings function and the way we get the longest common substring.

no task id

closes odoo/enterprise#91827

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>