[IMP] ai: support web grounding in LLMApiService
Purpose:
-------
Adds a parameter `web_grounding` to the `request_llm` method which, when
set to True, enables the LLM to perform web grounding by retrieving and
incorporating relevant information from the web into its response.

Updated ai_fields so that it uses the `_request_llm` method now that
web grounding is supported.

With this commit, system prompts won't be sent as 'agent' messages anymore.
Instead, they will be sent as a `systemInstruction`.

Task-4989962

Part-of: odoo/enterprise#91842
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>