[FIX] sign: fix drag and drop step in sign_tour
The `sign_tour` was failing after the conversion of the `sign.Template`
client action to OWL. The standard `drag_and_drop` tour helper can no
longer be used for automatic tour testing because the drop target is
inside an iframe whose content is managed by PDF.js.

This commit fixes the tour by making use of the existing custom helper
method, `dragAndDropSignItemAtHeight`, to programmatically simulate
the drag and drop action.

The helper method was introduced by following commit
https://github.com/odoo/enterprise/commit/664923afe8adfdab5b75957911d9acf9560145b3#diff-ea1fbf4d2936d0c3aa4b93cdf1f73d024a834159b350c34e4f383ef0106f0b81

For scalability, this commit also makes the helper method
importable & added the `pageNumber` parameter to the method.

opw-4752794

X-original-commit: aacd3e21d00f7cd8a17bd41f1be4ce97933ede3b
Part-of: odoo/enterprise#94420
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Sunhyeok Jung (suju) <suju@odoo.com>