[FIX] sign: do not copy sign items for sign_tour template.
If `sign.template_sign_tour` has sign request, it means that the
template might have a sign item because the `sign_tour` tour adds the
Signature sign item to the template.

Let's say another user tries the tour.
The tour then guide the user to add `Signature` sign item to the
template, ending up having two sign items; one from previous tour and
another from current tour. Tour will stall because there's more sign
items than it expects.

Skipping the `adding Signature sign item` step to tackle this is not
desired, as we want to guide the user how to add the sign item from side
bar to the sign template. ( Even though the UI is intuitive enough w/o
the guide ).

Therefore, when we're copying the sign template to trigger the template
tour, we should not copy the sign item.

opw-4752794

closes odoo/enterprise#94420

X-original-commit: e5a306b1efcd7a5420729e8893e48d9bfcb0b8d4
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Sunhyeok Jung (suju) <suju@odoo.com>