[FIX] l10n_lu_reports{,_annual_vat_2023}: Luxembourg tax report xml export special schemes
As per legal requirements, the 491, 492 and 493 fields of the xml export of the
tax return only need to be included if at least one day of the period includes
dates in 2025 or later.
As per legal requirements too, the 192 and 193 fields are mandatory in the
annual report if some other fields are present. Among those are 361 and 362
which are mandatory even if null so if 192 and 193 are not there yet, we set
them to 362 and 363 values (0.0 in most cases).

from feedback on
task-4587067
opw-4757770

closes odoo/enterprise#91966

X-original-commit: 38d43e0c4a10182eee4bedc83219107c530afbd0
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>