[IMP] sign: moved sign items rendering to an on-change basis
PDF.js, the library used for the rendering of the PDFs, sometimes tears
down the whole page and rerenders it from scratch. Because of this, we
have to add our custom sign fields on top periodically to avoid them
disappearing. Currently, we render all the fields every second but this
leaves a window between them disappearning and being rerendered which
shows as the fields flickering. With this PR we want to transition the
rerendering of the fields from a periodic nature to an event-based
nature by catching the event of the page being rerendered from the
eventBus of PDF.js

closes odoo/enterprise#91887

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>