[FIX] sale_renting_planning: make sure to not overload the rental service
Before this commit, when the user creates a new shift and clicks on `Add
to Last Order` the shift will be added to the SOL contained one of
products linked to the role set on the shift. The problem is we don't
check if there is still something to plan for that SOL.

This commit makes sure the SOL to set to the shift has something to plan
to be sure it is not overloaded.

task-4916537

closes odoo/enterprise#92260

X-original-commit: 0592a2ef477054b1741b14e26e3d9618584f3d0c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>