[IMP] ai: ask ai command available in default namespace
This commit allows access to the ask ai command (natural language query feature)
from the default namespace (CTRL+K) for better UX. After this commit, Ask AI
agent can now be access from both the "/" and default namespace.

To accomplish the feature, we need to override the `setCommands` of the CommandPalette
component because it has special handling of "default" commands -- it filters out
the commands not similar to the searchValue. However, we want the Ask AI command to be
always there (showing as well the searchValue). So, when we see that the set of filtered
commands has zero items, that's the time we insert the Ask AI command. We are also now
caching the Ask AI agent to avoid unnecessary rpc calls.

Part-of: odoo/enterprise#92870
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>