[IMP] web_studio: multiple improvements in studio
- Removing properties from chatter

- Added related field will take by default the string
of the related field

- Correction of the property 'Color field' for the
many2many_tags widget. The value wasn't showing.

- A new attribute is available for the supportedOptions.
It's 'isRelationalField'. This attribute is used for the
options that are type: "field". When it's set to true, the
list of fields for the selection will come from the relation
model and not from the viewEditor model.

- The techinal name is now normalized

- and other minor UX improvements

closes odoo/enterprise#92729

Task-id: 4936789
Related: odoo/odoo#223537
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>