[FIX] documents(,_spreadsheet): restore ‘export’ and ‘insert’ actions
Issue:
When clicking on the action dropdown in the kanban and list view,
the actions 'Export' and 'Insert in spreadsheet' did not appear
anymore (result of ).

Cause:
This was introduced by 8e7c92d1 that aimed to hide those
actions when a user is in the preview of a document.

Solution:
Adapt the omitting logic to only hide the actions when a document
is in preview.

Task-4898152

closes odoo/enterprise#92460

X-original-commit: 488a1e4e9aa704a39b4ac0a6a6d14bf0bb00ff70
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>