[FIX] l10n_fr_reports: fix tax report pdf inconsistency
When you validate a tax return, it will lock the date and generate the pdf for the tax report.

In France with the edi, if you have to send a reimbursement declaration, the remaining value of this reimbursement will be carryover for the next period, moving the value from line 27 to line 26.

With this change, the pdf of the report will change as well, but we still keep the old pdf in the return attachments, which mean we have 2 pdf with different values for the same tax period.

With this commit, we regenerate the pdf once the reimbursement declaration is sent, with the good values.

Also, if the user reset the account return, the external value originally created when validating is deleted.

task-4844276, task-4932427

Part-of: odoo/enterprise#93886
Related: odoo/upgrade#8379
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>