[FIX] l10n_mx_edi_sale: trigger compute qty invoiced cfdi state
How to reproduce the issue:
In l10n_mx:
1. Create a sale order.

2. Create an invoice from the SO, with yesterday’s date, and send it to the CFDI.

3. Lock the period at yesterday’s date.

4. Request the invoice cancellation with the “01” method. This creates a new invoice — confirm it.

5. Create a credit note for the original invoice.

6. Retry sending the cancel request to the CFDI; the CFDI state is now cancelled for the original invoice.

On the original SO, the quantity invoiced is 0 despite the CFDI state of the original invoice being cancelled.

This commit forces a recompute of the quantity invoiced when the CFDI state changes.
Since _compute_qty_invoiced uses _get_invoice_lines, which filters out invoices in sent state, only lines from invoices in cancelled state will be taken into account.

opw-4910139

closes odoo/enterprise#93582

X-original-commit: 2240cbc6db6e5fd36376e15be7ea89d3506493b8
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>