[FIX] sale_renting_planning: open form view of rental order in full page
This commit alters the way we open form view of a rental order once the
user would like to create a Rental Order from a planning slot. Now the
form view we will fully open, that is, it will no longer be opened in a
modal. The reason is because when the user when to create rental order
from a planning slot, it means the setting of planning slot is done and
can directly be redirected to the rental order to generate instead of
going back to the planning slot once the rental order is created.

task-4916650

Part-of: odoo/enterprise#93670
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>