[IMP] account_reports: add option to disable account report footer
When generating audit reports, we compile a single, comprehensive PDF
that includes all relevant documents and account reports attached to
the articles. Each page in the final report is automatically numbered
during generation.

However, because account reports also include their own page numbers,
this can lead to pages displaying two conflicting numbers: one from the
account report and another from the audit report.

To address this, we're introducing a new context variable in the account
reports (named: `exclude_page_footer`). This allows the page footer to
be disabled when needed. Disabling the account report's footer ensures
that the final audit report is cleaner, more consistent, and easier for
users to read.

Task-4989809

closes odoo/enterprise#91848

Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>