[IMP] l10n_co_dian: automate certification process
New customers starting out with odoo in Colombia have to go through a
registration process where they send specific records to the DIAN via a
specific endpoint. Currently, the process is:

1) Activate Certification process in settings.
2) Create 3 journals (SETP, NC, ND) for invoice, debit notes, and credit
   notes.
3) Create a specific product with a tax enabled.
4) Create the specific amount of invoices/debit notes/credit notes as
   required from within the DIAN portal.

This process is slow and for most of the users coming to Odoo, they will
not have experience with accounting software or understand where to
setup such records. As such, a one click setting has been added for
certification where they input the number of records necessary, and it
will automatically do everything they need.

Notes on implementation:
- This will perform changes that could be considered irreversible on
  customers databases. As such there is a warning that it should only
  occur on a staging/duplicated database.
- The values that are hard coded, Technical Key, Min/Max Range Values,
  and the Authorization Number all come from DIAN and are shared across
  all users, as such hard coding them to these values are not releasing
  any secrets.

  task-4366022

closes odoo/enterprise#91755

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>