[FIX] l10n_pe_reports: remove modification on move_name
Problem: For PLE reports 5.1 and 6.1, the value for `glosa` was
being modified (i.e. blank spaces and slashes removed). Although
SUNAT accepts the modified value, no information should be removed
in the case it might lead to issues with tax authorities during an
audit.

Solution: The value for `glosa` on the reports should be the exact
account move name.

opw-4803060

closes odoo/enterprise#92458

X-original-commit: 49f5f1126fce7d1140eab76ccbb2a7cc4830095d
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>