[IMP] documents_project{_sale}: copy all documents when duplicating project
Before this commit, when the user creates a project template and set it
to a product to generate a project based on that template when the SO
contained that product is confirmed, the problem is the files contained
the document folder of that project template are not copied when the
project template is duplicated. Same when the user duplicates a project.

This commit makes sure the files are duplicated when a project is duplicated.

task-5005032

Part-of: odoo/enterprise#91990
Related: odoo/odoo#221359
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>