[IMP] appointment: move to kanban view for gantt popover
Using a kanban view will have gantt instanciate a whole record properly
and load the kanban view. Which will in turn enable us to use widgets
normally inside that view and even use the model for value updates
instead of rolling our own.

The trade-off is that we cannot have ad-hoc values in there anymore so
we need to add a few fields for the "contact info"

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>