[IMP] appointment: remove badges slot counters on calendar chevrons
The small badges with the number of available slots in previous months
or next months, (supposed to) appear next to the chevrons on the calendar
in the booking flow, have a low interest and only work for 'custom'
appointments. As deemed disposable, we get rid of them and of the related
test.

Task-4997980

Part-of: odoo/enterprise#91824
Related: odoo/upgrade#8206
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>