[FIX] stock_fleet_enterprise: use correct map view in action
Commit ade5e5085df2 move the `default_order` value as it was defined in
the wrong module. Consequence is the map view defined in
`stock_enterprise` cannot be reordered anymore. This is an issue as the
one in `stock_fleet_enterprise` is never used. The action window is
patched to take a picking map view but without specifying witch one.

This commit force the map view defined in `stock_fleet_enterprise` to be
used in the action of `stock_fleet_enterprise`.

Consequence is that the picking can be rearranged again in the map view
coming from a batch transfer and not coming from the picking list.

closes odoo/enterprise#92045

X-original-commit: a7f2a2cf9c96939543abcd72a71c483d8e5c805f
Signed-off-by: William Henrotin (whe) <whe@odoo.com>