[IMP] l10n_be_reports: Replace TXT export with XML for Annual Accounts
The previous Annual Accounts export in the General Ledger generated a TXT file with a fixed column order, which did not meet the structural
or formatting requirements for downstream usage.

This commit replaces the TXT export with an XML export conforming to the official format, including fields such as account code, debit,
credit, multilingual descriptions, and metadata.

task-5005345

closes odoo/enterprise#92588

X-original-commit: 1bd313feb1071bf735dfb691830cc04731929fe2
Signed-off-by: John Laterre (jol) <jol@odoo.com>