[IMP] appointment,web_gantt: reuse the calendar quick-create dialog
Calendar has a custom dialog for quick create that enables opening a new
form view with defaults from the quick create. This enables users to
finalize the details of the appointment without notifying them of the
creation, which could lead to multiple notifications in quick
succession.

Gantt is slightly modified to allow opening the quick-edit dialog with
different dialog components and additional porps.

Also add "total capacity" to the form view so it can be propagated
from the "more options" button. And more generally to help see the
actual used capacity at a glance.

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>