[IMP] l10n_ke_hr_payroll: add tier 3 to nssf report and option to report
In Kenya, you can report your pension contribution to the government (NSSF
Report). So we add TIER 3, with code 103, to the report which is for pension.

We also add the possibility to choose for TIER 2 and TIER 3 if we want to
have it on the NSSF Report or if it's given to a third party insurance company.

closes odoo/enterprise#92024

Task: 5003604
Related: odoo/upgrade#8204
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>